Add basic suite mirroring without processing
Adds a Robot framework compatible keyword library with a single keyword. When used in the setup of a test suite it converts that test suite into a data structure and replaces the contents of that test suite with a generated one. Without any further processing, there should be no difference between the initial and the final test suites.