First basic implementation that allow repetition
Note that this commit has regression on the existing scenarios, does not satisfy the new scenarios, nor can it detect the impossible trace. It can however already handle all scenarios with repetition if the randomiser gets lucky.

All failures are due to the fact that the algorithm does not know when to stop trying.